Convert the output file to csv format
The data files generated in the SD card with each restart were formatted
in a way that was difficult to convert into a data frame for post
processing. There were blank columns and extra spaces that needed to be
removed. The first line in the data was supposed to be a timestamp but
that function doesn't currently work and makes the output unecessarily
messy.

I changed the library inclusion for SDLogger.h from <> to "" because
that was how I was able to access the library in Arduino, but this may
need to be changed back.

Resolves: #20