Output more precise Hz and fix random crashing
It takes about 5-6 ms for the MKR to read the IMU reading which
makes the sampling rate closer to 18Hz instead of 20. To account
for this, the delay between IMU reads was changed to 45ms instead
of 50ms which puts all the readings closer to the 20Hz mark.

The size of the buffer string that stores the readings before they
are written to the SD card was too large which was causing
unexpected output and crashing during data collection. The
buffer size was decreased which fixed the unexpected output
but the downside to this solution is that there are more SD
card writes which disrupts the precision of the readings.

Resolves: #24