Add script for processing raw WitMotion data files
Appeared to work on example files that Katie added to repo.  Should be
tested further.  Timestamp decoding (and inclusion in the CSV) not yet
implemented. There are some assumptions in the code about the data
structure that might not hold up with other WitMotion data so testing
with other data also warranted.